tests: moved config fixture
