+++ tests/test_unit/conftest.py
@@ -8,14 +8,17 @@ from typing import Any
 
 import pytest
 from cognite.client import global_config
+from cognite.client.credentials import Token
 from cognite.client.data_classes import CreatedSession
 from pytest import MonkeyPatch
 
+from cognite_toolkit._cdf_tk.client import ToolkitClientConfig
 from cognite_toolkit._cdf_tk.client.testing import monkeypatch_toolkit_client
 from cognite_toolkit._cdf_tk.commands import ModulesCommand, RepoCommand
 from cognite_toolkit._cdf_tk.utils.auth import EnvironmentVariables
 from tests.constants import REPO_ROOT
 from tests.test_unit.approval_client import ApprovalToolkitClient
+from tests.test_unit.test_cdf_tk.test_client.conftest import BASE_URL
 from tests.test_unit.utils import PrintCapture
 
 THIS_FOLDER = Path(__file__).resolve().parent
@@ -140,3 +143,15 @@ def disable_gzip():
     global_config.disable_gzip = True
     yield
     global_config.disable_gzip = old
+
+
+@pytest.fixture
+def toolkit_config():
+    return ToolkitClientConfig(
+        client_name="test-client",
+        project="test-project",
+        base_url=BASE_URL,
+        max_workers=1,
+        timeout=10,
+        credentials=Token("abc"),
+    )
